Fixed numerous typos. Fixed a bug that prevented physical connections to be displayed correctly. Container and Link positions in SDH module are casted correctly. Scripts have now .groovy extension

+++ client/trunk/ContractManagerModule/src/com/neotropic/inventory/modules/contracts/nodes/actions/DeleteContractAction.java
@@ -20,6 +20,7 @@ import com.neotropic.inventory.modules.contracts.nodes.ContractPoolNode;
 import java.awt.event.ActionEvent;
 import java.util.Iterator;
 import javax.swing.AbstractAction;
+import javax.swing.JOptionPane;
 import org.inventory.communications.CommunicationsStub;
 import org.inventory.core.services.api.notifications.NotificationUtil;
 import org.openide.util.Utilities;
@@ -36,19 +37,23 @@ public class DeleteContractAction extends AbstractAction {
     
     @Override
     public void actionPerformed(ActionEvent e) {
-        Iterator<? extends ContractNode> selectedNodes = Utilities.actionsGlobalContext().lookupResult(ContractNode.class).allInstances().iterator();
-            
-        if (!selectedNodes.hasNext())
-            return;
-    
-        ContractNode selectedNode = selectedNodes.next();
+        if (JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this contract?", 
+                "Warning", JOptionPane.OK_CANCEL_OPTION) == JOptionPane.OK_OPTION) {
         
-        if (CommunicationsStub.getInstance().deleteObject(selectedNode.getObject().getClassName(), selectedNode.getObject().getOid())) {
-            NotificationUtil.getInstance().showSimplePopup("Information", NotificationUtil.INFO_MESSAGE, "The selected contract was deleted");
-            ((ContractPoolNode.ContractPoolChildren)selectedNode.getParentNode().getChildren()).addNotify();
+            Iterator<? extends ContractNode> selectedNodes = Utilities.actionsGlobalContext().lookupResult(ContractNode.class).allInstances().iterator();
+
+            if (!selectedNodes.hasNext())
+                return;
+
+            ContractNode selectedNode = selectedNodes.next();
+
+            if (CommunicationsStub.getInstance().deleteObject(selectedNode.getObject().getClassName(), selectedNode.getObject().getOid())) {
+                NotificationUtil.getInstance().showSimplePopup("Information", NotificationUtil.INFO_MESSAGE, "The selected contract was deleted");
+                ((ContractPoolNode.ContractPoolChildren)selectedNode.getParentNode().getChildren()).addNotify();
+            }
+            else
+                NotificationUtil.getInstance().showSimplePopup("Error", NotificationUtil.INFO_MESSAGE, CommunicationsStub.getInstance().getError());
         }
-        else
-            NotificationUtil.getInstance().showSimplePopup("Error", NotificationUtil.INFO_MESSAGE, CommunicationsStub.getInstance().getError());
             
     }
